prototypes/hpc-runner/pyproject.toml
[project]
name = "hpc-regression"
version = "0.1.0"
description = "Minimal runner harness using structlog"
requires-python = ">=3.10"
dependencies = [
  "structlog",
  "pyyaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["hpc_regression"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = ["pytest"]

